# sisyphus-server/tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sisyphus Server API",
    "description": "自托管 CI 平台 Server REST API。业务端点均在 /api/v1/ 前缀下；B2b 起会话认证全局把关（未认证一律 401，放行面仅 login/setup/healthz 与静态资源，票 B2b-T1）。",
    "contact": {
      "name": "sisyphus contributors"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/agent/upgrade-packages/{package_name}": {
      "get": {
        "tags": [
          "upgrade-packages"
        ],
        "summary": "下载升级包（Agent token 鉴权；Agent 侧拉取后校验 sha256、原子换入）。",
        "operationId": "download",
        "parameters": [
          {
            "name": "package_name",
            "in": "path",
            "description": "包名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "包字节流（响应头带 size/sha256）"
          },
          "401": {
            "description": "Agent token 无效/已停用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "包不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Agent 清单（全局 admin；按名排序，含已停用，无凭据值形态）。",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "全部 Agent（按名排序，含已停用）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "非全局管理员",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "建 Agent 条目（全局 admin）：签发 token + 注册码，落库行（离线、未\n停用；系统标签空——由首次连接上报）。",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "已创建；token 与注册码明文仅此一次返回",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAgentResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "非全局管理员",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Agent 名已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "输入校验失败（名非空/字符集、标签形态、槽位 >= 1）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{name}/pipelines/{pipeline}/builds/{number}/artifacts/{artifact}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "单产物下载（viewer 档，票 #74）：流式响应，响应头带大小\n（Content-Length）与校验和（X-Sisyphus-Sha256）。",
        "operationId": "download",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "项目名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline",
            "in": "path",
            "description": "pipeline 名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "description": "构建号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "description": "产物名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "产物字节流（Content-Length = 大小，X-Sisyphus-Sha256 = 校验和）",
            "content": {
              "application/octet-stream": {}
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "权限不足（需 viewer 档）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "项目不存在/不可见，构建号或产物不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{name}/pipelines/{pipeline}/builds/{number}/jobs/{job}/attempts/{attempt}/logs": {
      "get": {
        "tags": [
          "builds"
        ],
        "summary": "整份日志下载端点（viewer 档，ADR-0013）：text/plain，全部 chunk 解压\n拼接为纯文本（输出原样含 ANSI、步骤回显、截断标注）。",
        "operationId": "download",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "项目名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline",
            "in": "path",
            "description": "pipeline 名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "in": "path",
            "description": "构建号",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "job",
            "in": "path",
            "description": "任务名",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "description": "第几次执行",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "整份日志纯文本（text/plain；输出原样含 ANSI、步骤回显 $ 命令）",
            "content": {
              "text/plain": {}
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "权限不足（下载日志需 viewer 档）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "项目不存在/不可见，或构建号/任务/attempt 不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/upgrade-packages": {
      "get": {
        "tags": [
          "upgrade-packages"
        ],
        "summary": "升级包清单（全局 admin；按包名排序）。",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "全部升级包（按包名排序）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UpgradePackageResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "upgrade-packages"
        ],
        "summary": "上传升级包（全局 admin）：raw octet-stream body + `X-Sisyphus-Filename` 头。",
        "operationId": "upload",
        "responses": {
          "201": {
            "description": "已上传；返回包元数据（含解析版本/目标/sha256）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpgradePackageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "版本窗外（< N-1 或 > Server）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "文件名缺/不可解析/目标三元组非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/upgrade-packages/{package_name}": {
      "delete": {
        "tags": [
          "upgrade-packages"
        ],
        "summary": "删除升级包（全局 admin）：删元数据行 + 字节文件。",
        "operationId": "delete",
        "parameters": [
          {
            "name": "package_name",
            "in": "path",
            "description": "包名",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "已删除"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "包不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentResponse": {
        "type": "object",
        "description": "Agent 管理视图（无 token/注册码值形态——明文只在创建响应出现一次）。",
        "required": [
          "name",
          "online",
          "disabled",
          "system_labels",
          "custom_labels",
          "max_concurrency",
          "active_jobs",
          "version_compatible",
          "draining",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "active_jobs": {
            "type": "integer",
            "format": "int64",
            "description": "在途任务数（running/unknown 占槽，ADR-0008 中心化计数）。"
          },
          "agent_version": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VersionDto",
                "description": "握手上报的 Agent 版本（ADR-0017；从未握手为空）。"
              }
            ]
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（Unix 毫秒）。"
          },
          "custom_labels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "自定义标签（管理员可编辑）。"
          },
          "disabled": {
            "type": "boolean",
            "description": "停用标志（停用即踢线）。"
          },
          "disk_usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DiskUsageDto",
                "description": "磁盘占用（ADR-0019：卷级/缓存/工作区最近采样；从未上报为空）。"
              }
            ]
          },
          "draining": {
            "type": "boolean",
            "description": "排空/升级中（在线但不可派发——pending 升级指令或 draining/downloading/\nswapping/restarting 阶段；fallback 与无升级为 false，ADR-0017）。"
          },
          "last_seen_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "最近心跳时间（Unix 毫秒；从未在线为空）。"
          },
          "max_concurrency": {
            "type": "integer",
            "format": "int32",
            "description": "并发槽位数。"
          },
          "name": {
            "type": "string",
            "description": "构建机名（唯一）。"
          },
          "online": {
            "type": "boolean",
            "description": "在线状态（心跳 15s 收、45s 无心跳判离线）。"
          },
          "system_labels": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "系统事实标签（sisyphus/os、sisyphus/arch、sisyphus/container；由\n注册/心跳上报，不可手编）。"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "最后更新时间（Unix 毫秒）。"
          },
          "upgrade_error": {
            "type": [
              "string",
              "null"
            ],
            "description": "升级失败原因（fallback 时记；否则为空）。"
          },
          "upgrade_phase": {
            "type": [
              "string",
              "null"
            ],
            "description": "升级阶段（draining/downloading/swapping/restarting/fallback；无升级为空）。"
          },
          "version_compatible": {
            "type": "boolean",
            "description": "版本是否兼容（落在 N-1 窗口内；过旧/过新为 false，ADR-0017 四态派生面）。"
          }
        }
      },
...
      "CreateAgentRequest": {
        "type": "object",
        "description": "建 Agent 条目请求体（全局 admin）。",
        "required": [
          "name"
        ],
        "properties": {
          "custom_labels": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "自定义标签（key=value 字符串数组，可空/省略 = 空集；管理员可编辑，\n匹配与系统标签取并集做 AND 全集语义）。"
          },
          "max_concurrency": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "并发槽位数（默认 1；>= 1）。"
          },
          "name": {
            "type": "string",
            "description": "构建机名（唯一；1-64 位字母数字或 `_ . -`，trim 后生效——名会进\nURL 路径 `/agents/{name}`）。"
          }
        }
      },
...
      "CreatedAgentResponse": {
        "type": "object",
        "description": "建条目响应：token 与注册码明文仅此一次返回（此后任何端点不再出现值；\n库里只存各自哈希，任何端点都无法找回）。",
        "required": [
          "token",
          "register_code",
          "agent"
        ],
        "properties": {
          "agent": {
            "$ref": "#/components/schemas/AgentResponse",
            "description": "Agent 管理视图。"
          },
          "register_code": {
            "type": "string",
            "description": "一次性注册码（`sisa_reg_` + 43 字符；注册码换 token 流程随 Agent\n批次，本批建条目即签发 token）。"
          },
          "token": {
            "type": "string",
            "description": "per-Agent 通道 token（`sisa_` + 43 字符）。请立即保存：本响应是\n唯一一次出现。"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "统一错误响应体。",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码（大写蛇形，如 `NOT_FOUND`、`VALIDATION_FAILED`）。"
          },
          "detail": {
            "description": "结构化补充（如校验错误清单）；无补充时缺省不输出。"
          },
          "message": {
            "type": "string",
            "description": "人读错误信息，可直接展示。"
          }
        }
      },
...
      "UpgradePackageResponse": {
        "type": "object",
        "description": "升级包视图（与 [`crate::store::upgrade_packages::UpgradePackageMeta`] 同构）。",
        "required": [
          "package_name",
          "version",
          "target_os",
          "target_arch",
          "size",
          "sha256",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "上传时刻（Unix 毫秒）。"
          },
          "package_name": {
            "type": "string",
            "description": "包名（ADR-0010 规范 `sisyphus-agent-<ver>-<os>-<arch>`）。"
          },
          "sha256": {
            "type": "string",
            "description": "SHA-256 校验和（十六进制小写）。"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "字节数。",
            "minimum": 0
          },
          "target_arch": {
            "type": "string",
            "description": "目标架构（x86_64/aarch64）。"
          },
          "target_os": {
            "type": "string",
            "description": "目标 OS（linux/macos/windows）。"
          },
          "version": {
            "$ref": "#/components/schemas/VersionDto",
            "description": "解析自文件名的版本。"
          }
        }
      },
...
    }
  },
...
}